Add useJSONData hook Move common data fetching code into a custom hook for JSON data subjects.
