+++ packages/lib-classifier/src/helpers/index.js
@@ -8,6 +8,7 @@ export { default as layouts } from './layouts'
 export { default as shownMarks } from './shownMarks'
 export { default as subjectsSeenThisSession } from './subjectsSeenThisSession'
 export { default as subjectViewers } from './subjectViewers'
+export { default as useJSONData } from './useJSONData'
 export { default as useStores } from './useStores'
 export { default as validateMimeType } from './validateMimeType'
 export { default as validateSubjectLocations } from './validateSubjectLocations'
