+++ packages/lib-classifier/src/helpers/useJSONData/README.md
@@ -1,0 +1,17 @@
+# useJSONData
+  
+A custom hook which loads JSON data for a Panoptes subject.
+
+Usage:
+```js
+import { useJSONData } from '@helpers'
+
+
+function MyDataComponent({ subject, onReady, onError }) {
+  const JSONdata = useJSONData(
+    subject,
+    () => onReady(),
+    (error) => onError(error)
+  )
+}
+```
