+++ packages/lib-classifier/src/helpers/useJSONData/index.js
@@ -1,0 +1,1 @@
+export { default } from './useJSONData'
